+++ src/main/java/com/yfkyplatform/parkinglotmiddleware/carpark/daoer/client/domin/api/IDaoerCarPort.java
@@ -4,6 +4,7 @@ import com.yfkyplatform.parkinglotmiddleware.carpark.daoer.client.domin.resp.car
 import com.yfkyplatform.parkinglotmiddleware.carpark.daoer.client.domin.resp.daoerbase.DaoerBaseResp;
 import reactor.core.publisher.Mono;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -25,21 +26,23 @@ public interface IDaoerCarPort {
      * @return
      */
     Mono<DaoerBaseResp<CarFeeResult>> getCarFeeInfo(String carNo);
+
     /**
      * 临停缴费支付完成
      *
      * @param carNo 车牌号码
      * @return
      */
-    Mono<DaoerBaseResp> payCarFeeAccess(String carNo, String payTime, int duration, double totalAmount, double disAmount,
-                                               int paymentType,int payType,String paymentTnx,double couponAmount,String channelId);
+    Mono<DaoerBaseResp> payCarFeeAccess(String carNo, String payTime, int duration, BigDecimal totalAmount, BigDecimal disAmount,
+                                        int paymentType, int payType, String paymentTnx, BigDecimal couponAmount, String channelId);
+
     /**
      * 根据通道号获取车辆费用信息
      *
      * @param carNo 车牌号码
      * @return
      */
-    Mono<DaoerBaseResp<CarFeeResult>> getChannelCarFee(String channelId,String carNo, String openId);
+    Mono<DaoerBaseResp<CarFeeResult>> getChannelCarFee(String channelId, String carNo, String openId);
     /**
      * 无牌车入场
      *
